refactor: move Footer component to base-gallery module
