+++ src/components/base-gallery/Footer.tsx
